Add Travis CI automated testing
The owner of the this repo would need to go to https://travis-ci.com/profile, log in with GitHub credentials, ___select the free plan___, and flip the repository switch on to enable free automated testing of all pull requests.  Travis will lint with [flake8](http://flake8.pycqa.org) and then run __python build.py__.  Python 2 and Python 3 tests run in parallel but the Python 3 tests are run in __allow_failures__ mode until the port is complete.  Travis CI is free to open source project like this one.