Run reindent.py winpwnage.py
This commit is the result of running __reindent.py winpwnage.py__ where ___reindent.py___ is a file that is shipped in every CPython distribution.
* https://github.com/python/cpython/blob/master/Tools/scripts/reindent.py

Without this change PyInstaller builds succeed but the resulting .exe files generate nothing on stdout.  This is probably caused by a bare exception somewhere in the codebase.